fixed: ensure  unsigned values for the calculation
svn path=/trunk/KDE/kdeutils/okteta/; revision=968067

+++ program/oktetakakao/controllers/view/libbytearraychecksum/algorithm/modsum16bytearraychecksumalgorithm.cpp
@@ -44,10 +44,10 @@ bool ModSum16ByteArrayChecksumAlgorithm::calculateChecksum( QString* result,
     // TODO: move padding checks into extra code before and after loop
     for( int i = section.start(); i<=section.end(); ++i )
     {
-        int value = model->datum( i ) << 8;
+        quint16 value = (quint8)( model->datum(i) ) << 8;
         ++i;
         if( i<=section.end() )
-            value += model->datum( i );
+            value += (quint8)( model->datum(i) );
 
         modSum += value;
 #if 0
